refactor(daytona): architectural deepening of integrations/daytona layer
- Add protocols.py: runtime-checkable Protocols (DaytonaSandbox, DaytonaClient,
  sub-service Protocols) replacing Any at SDK seam boundaries (19 annotations typed)
- Decompose sdk_ops.py (34KB) into focused modules:
  volumes.py (volume readiness/layout/listing)
  snapshots.py (snapshot CRUD/bootstrap/resolution/fallback)
  file_browser.py (volume tree/file preview)
  sdk_ops.py retained as slim barrel + sandbox lifecycle helpers
- Split workspace_runtime.py into:
  _git_helpers.py (25 pure git helper functions, 501 lines)
  workspace_runtime.py (session orchestration + Pydantic models, 218 lines)
- Deepen DaytonaSandboxRuntime: inline sandbox creation logic from sandbox_lifecycle
  into class methods; eliminate sandbox_lifecycle.py intermediary
- Apply async-only canonical convention to internal modules: sync names removed from
  public APIs of volumes/snapshots/file_browser; DaytonaSandboxRuntime public sync API preserved

All 288 unit tests pass. format-check/lint/typecheck clean.
